moved sql and node info related functions to separate files fixed some logging messages variables renamed
