+++ p2p/server.go
@@ -22,18 +22,13 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"math"
-	"math/big"
 	"net"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/teamnsrg/go-ethereum/common"
 	"github.com/teamnsrg/go-ethereum/common/mclock"
-	"github.com/teamnsrg/go-ethereum/crypto"
 	"github.com/teamnsrg/go-ethereum/event"
 	"github.com/teamnsrg/go-ethereum/log"
 	"github.com/teamnsrg/go-ethereum/p2p/discover"
@@ -356,45 +351,6 @@ func (srv *Server) makeSelf(listener net.Listener, ntab discoverTable) *discover
 	return ntab.Self()
 }
 
-func (srv *Server) CloseSQL() {
-	if srv.DB != nil {
-		if srv.addNodeInfoStmt != nil {
-			if err := srv.addNodeInfoStmt.Close(); err != nil {
-				log.Debug("Failed to close AddNodeInfo sql statement", "err", err)
-			} else {
-				log.Trace("Closed AddNodeInfo sql statement")
-			}
-		}
-		if srv.updateNodeInfoStmt != nil {
-			if err := srv.updateNodeInfoStmt.Close(); err != nil {
-				log.Debug("Failed to close UpdateNodeInfo sql statement", "err", err)
-			} else {
-				log.Trace("Closed UpdateNodeInfo sql statement")
-			}
-		}
-		if srv.addNodeMetaInfoStmt != nil {
-			if err := srv.addNodeMetaInfoStmt.Close(); err != nil {
-				log.Debug("Failed to close AddNodeMetaInfo sql statement", "err", err)
-			} else {
-				log.Trace("Closed AddNodeMetaInfo sql statement")
-			}
-		}
-		if srv.GetRowIDStmt != nil {
-			if err := srv.GetRowIDStmt.Close(); err != nil {
-				log.Debug("Failed to close GetRowID sql statement", "err", err)
-			} else {
-				log.Trace("Closed GetRowID sql statement")
-			}
-		}
-		driver := "mysql"
-		if err := srv.DB.Close(); err != nil {
-			log.Debug("Failed to close sql db handle", "database", srv.MySQLName, "driver", driver, "err", err)
-		} else {
-			log.Trace("Closed sql db handle", "database", srv.MySQLName, "driver", driver)
-		}
-	}
-}
-
 // Stop terminates the server and all active peer connections.
 // It blocks until all active connections have been closed.
 func (srv *Server) Stop() {
@@ -411,8 +367,7 @@ func (srv *Server) Stop() {
 	close(srv.quit)
 	srv.loopWG.Wait()
 
-	// close mysql db handle
-	srv.CloseSQL()
+	srv.CloseSql()
 }
 
 // Start starts running the server.
@@ -424,43 +379,12 @@ func (srv *Server) Start() (err error) {
 		return errors.New("server already running")
 	}
 
-	// open mysql db handle
-	if srv.MySQLName != "" {
-		driver := "mysql"
-		db, err := sql.Open(driver, srv.MySQLName)
-		if err != nil {
-			log.Debug("Failed to open sql db handle", "database", srv.MySQLName, "driver", driver, "err", err)
-			return err
-		}
-		log.Trace("Opened sql db handle", "database", srv.MySQLName, "driver", driver)
-		err = db.Ping()
-		if err != nil {
-			log.Debug("Sql db connection failed ping test", "database", srv.MySQLName, "driver", driver, "err", err)
-			if err := srv.DB.Close(); err != nil {
-				log.Debug("Failed to close sql db handle", "database", srv.MySQLName, "driver", driver, "err", err)
-			} else {
-				log.Trace("Closed sql db handle", "database", srv.MySQLName, "driver", driver)
-			}
-			return err
-		}
-		log.Trace("Sql db connection passed ping test", "database", srv.MySQLName, "driver", driver)
-		srv.DB = db
-	}
-
 	srv.KnownNodeInfos = &knownNodeInfos{infos: make(map[discover.NodeID]*Info)}
 
-	if srv.DB != nil {
-		// fill KnownNodesInfos with info from the mysql database
-		srv.loadKnownNodeInfos()
-
-		// prepare sql statements
-		srv.prepareAddNodeInfoStmt()
-		srv.prepareUpdateNodeInfoStmt()
-		srv.prepareAddNodeMetaInfoStmt()
-		srv.prepareGetRowID()
+	// initiate sql connection and prepare statements
+	if err := srv.initSql(); err != nil {
+		return err
 	}
-
-	// TODO: load info from mysql db
 
 	srv.running = true
 	log.Info("Starting P2P networking")
@@ -980,426 +904,6 @@ func (srv *Server) NodeInfo() *NodeInfo {
 	return info
 }
 
-func (srv *Server) loadKnownNodeInfos() {
-	fields := "ni.id, ni.node_id, nmi.hash, ip, tcp_port, remote_port, " +
-		"p2p_version, client_id, caps, listen_port, first_hello_at, last_hello_at, " +
-		"protocol_version, network_id, first_received_td, last_received_td, best_hash, genesis_hash, " +
-		"first_status_at, last_status_at, dao_fork"
-	maxIds := "SELECT node_id as nid, MAX(id) as max_id FROM node_info GROUP BY node_id"
-	nodeInfos := fmt.Sprintf("SELECT * FROM node_info x INNER JOIN (%s) max_ids ON x.id = max_ids.max_id", maxIds)
-	stmt := fmt.Sprintf("SELECT %s FROM (%s) ni INNER JOIN node_meta_info nmi ON ni.node_id=nmi.node_id", fields, nodeInfos)
-	rows, _ := srv.DB.Query(stmt)
-
-	type sqlObjects struct {
-		p2pVersion      sql.NullInt64
-		clientId        sql.NullString
-		caps            sql.NullString
-		listenPort      sql.NullInt64
-		firstHelloAt    sql.NullFloat64
-		lastHelloAt     sql.NullFloat64
-		protocolVersion sql.NullInt64
-		networkId       sql.NullInt64
-		firstReceivedTd sql.NullString
-		lastReceivedTd  sql.NullString
-		bestHash        sql.NullString
-		genesisHash     sql.NullString
-		daoForkSupport  sql.NullInt64
-		firstStatusAt   sql.NullFloat64
-		lastStatusAt    sql.NullFloat64
-	}
-
-	srv.KnownNodeInfos.Lock()
-	defer srv.KnownNodeInfos.Unlock()
-
-	for rows.Next() {
-		var (
-			rowid      uint64
-			nodeid     string
-			hash       string
-			ip         string
-			tcpPort    uint16
-			remotePort uint16
-			sqlObj     sqlObjects
-		)
-		err := rows.Scan(&rowid, &nodeid, &hash, &ip, &tcpPort, &remotePort,
-			&sqlObj.p2pVersion, &sqlObj.clientId, &sqlObj.caps, &sqlObj.listenPort,
-			&sqlObj.firstHelloAt, &sqlObj.lastHelloAt, &sqlObj.protocolVersion, &sqlObj.networkId,
-			&sqlObj.firstReceivedTd, &sqlObj.lastReceivedTd, &sqlObj.bestHash, &sqlObj.genesisHash,
-			&sqlObj.daoForkSupport, &sqlObj.firstStatusAt, &sqlObj.lastStatusAt)
-		if err != nil {
-			log.Debug("MYSQL", "action", "query node info", "result", "fail", "err", err)
-			continue
-		}
-		// convert hex to NodeID
-		id, err := discover.HexID(nodeid)
-		if err != nil {
-			log.Debug("LOAD_FROM_MYSQL", "action", "parse node_id", "result", "fail", "err", err)
-			continue
-		}
-		nodeInfo := &Info{
-			RowID:         rowid,
-			Keccak256Hash: hash,
-			IP:            ip,
-			TCPPort:       tcpPort,
-			RemotePort:    remotePort,
-		}
-		if sqlObj.p2pVersion.Valid {
-			nodeInfo.P2PVersion = uint64(sqlObj.p2pVersion.Int64)
-		}
-		if sqlObj.clientId.Valid {
-			nodeInfo.ClientId = sqlObj.clientId.String
-		}
-		if sqlObj.caps.Valid {
-			nodeInfo.Caps = sqlObj.caps.String
-		}
-		if sqlObj.listenPort.Valid {
-			nodeInfo.ListenPort = uint16(sqlObj.listenPort.Int64)
-		}
-		if sqlObj.firstHelloAt.Valid {
-			i, f := math.Modf(sqlObj.firstHelloAt.Float64)
-			t := time.Unix(int64(i), int64(f*1000000000))
-			nodeInfo.FirstHelloAt = &t
-		}
-		if sqlObj.lastHelloAt.Valid {
-			i, f := math.Modf(sqlObj.lastHelloAt.Float64)
-			t := time.Unix(int64(i), int64(f*1000000000))
-			nodeInfo.LastHelloAt = &t
-		}
-		if sqlObj.protocolVersion.Valid {
-			nodeInfo.ProtocolVersion = uint64(sqlObj.protocolVersion.Int64)
-		}
-		if sqlObj.networkId.Valid {
-			nodeInfo.NetworkId = uint64(sqlObj.networkId.Int64)
-		}
-		if sqlObj.firstReceivedTd.Valid {
-			firstReceivedTd := &big.Int{}
-			s := sqlObj.firstReceivedTd.String
-			_, ok := firstReceivedTd.SetString(s, 10)
-			if !ok {
-				log.Debug("LOAD_FROM_MYSQL", "action", "parse *big.Int first_received_td", "result", "fail", "value", s)
-			} else {
-				nodeInfo.FirstReceivedTd = firstReceivedTd
-			}
-		}
-		if sqlObj.lastReceivedTd.Valid {
-			lastReceivedTd := &big.Int{}
-			s := sqlObj.lastReceivedTd.String
-			_, ok := lastReceivedTd.SetString(s, 10)
-			if !ok {
-				log.Debug("LOAD_FROM_MYSQL", "action", "parse *big.Int last_received_td", "result", "fail", "value", s)
-			} else {
-				nodeInfo.LastReceivedTd = lastReceivedTd
-			}
-		}
-		if sqlObj.bestHash.Valid {
-			nodeInfo.BestHash = sqlObj.bestHash.String
-		}
-		if sqlObj.genesisHash.Valid {
-			nodeInfo.GenesisHash = sqlObj.genesisHash.String
-		}
-		if sqlObj.firstStatusAt.Valid {
-			i, f := math.Modf(sqlObj.firstStatusAt.Float64)
-			t := time.Unix(int64(i), int64(f*1000000000))
-			nodeInfo.FirstStatusAt = &t
-		}
-		if sqlObj.lastStatusAt.Valid {
-			i, f := math.Modf(sqlObj.lastStatusAt.Float64)
-			t := time.Unix(int64(i), int64(f*1000000000))
-			nodeInfo.LastStatusAt = &t
-		}
-		if sqlObj.daoForkSupport.Valid {
-			var daoForkSupport bool
-			if uint16(sqlObj.daoForkSupport.Int64) != 0 {
-				daoForkSupport = true
-			}
-			nodeInfo.DAOForkSupport = daoForkSupport
-		}
-		srv.KnownNodeInfos.Infos()[id] = nodeInfo
-
-		// add the node to the initial static node list
-		srv.addInitialStatic(id, nodeInfo)
-	}
-}
-
-func (srv *Server) getNodeAddress(c *conn, receivedAt *time.Time) (*Info, bool, bool) {
-	var (
-		remoteIP   string
-		remotePort uint16
-		tcpPort    uint16
-		dial       bool
-		accept     bool
-	)
-	addrArr := strings.Split(c.fd.RemoteAddr().String(), ":")
-	addrLen := len(addrArr)
-	remoteIP = strings.Join(addrArr[:addrLen-1], ":")
-	if p, err := strconv.ParseUint(addrArr[addrLen-1], 10, 16); err == nil {
-		remotePort = uint16(p)
-	}
-	oldNodeInfo := srv.KnownNodeInfos.Infos()[c.id]
-
-	var hash string
-	if oldNodeInfo != nil {
-		hash = oldNodeInfo.Keccak256Hash
-		tcpPort = oldNodeInfo.TCPPort
-	} else {
-		hash = crypto.Keccak256Hash(c.id[:]).String()[2:]
-	}
-	// if inbound connection, resolve the node's listening port
-	// otherwise, remotePort is the listening port
-	if c.flags&inboundConn != 0 || c.flags&trustedConn != 0 {
-		if tcpPort == 0 {
-			newNode := srv.ntab.Resolve(c.id)
-			// if the node address is resolved, set the tcpPort
-			// otherwise, leave it as 0
-			if newNode != nil {
-				tcpPort = newNode.TCP
-			}
-		}
-		accept = true
-	} else {
-		tcpPort = remotePort
-		dial = true
-	}
-	newNodeInfo := &Info{
-		Keccak256Hash: hash,
-		LastHelloAt:   receivedAt,
-		IP:            remoteIP,
-		TCPPort:       tcpPort,
-		RemotePort:    remotePort,
-	}
-	return newNodeInfo, dial, accept
-}
-
-func (srv *Server) storeNodeInfo(c *conn, receivedAt *time.Time, hs *protoHandshake) {
-	// node address currentInfo
-	newInfo, dial, accept := srv.getNodeAddress(c, receivedAt)
-	id := hs.ID
-	nodeid := id.String()
-	if srv.addNodeMetaInfoStmt != nil {
-		srv.addNodeMetaInfo(nodeid, newInfo.Keccak256Hash, dial, accept, false)
-	}
-
-	// DEVp2p Hello
-	p2pVersion, clientId, capsArray, listenPort := hs.Version, hs.Name, hs.Caps, uint16(hs.ListenPort)
-	caps := ""
-	capsLen := len(capsArray)
-	for i, c := range capsArray {
-		caps += fmt.Sprintf("%s", c.String())
-		if i < capsLen-1 {
-			caps += ","
-		}
-	}
-	clientId = strings.Replace(clientId, "'", "", -1)
-	clientId = strings.Replace(clientId, "\"", "", -1)
-	caps = strings.Replace(caps, "'", "", -1)
-	caps = strings.Replace(caps, "\"", "", -1)
-
-	newInfo.P2PVersion = p2pVersion
-	newInfo.ClientId = clientId
-	newInfo.Caps = caps
-	newInfo.ListenPort = listenPort
-
-	srv.KnownNodeInfos.Lock()
-	defer srv.KnownNodeInfos.Unlock()
-	if currentInfo, ok := srv.KnownNodeInfos.Infos()[id]; !ok {
-		newInfo.FirstHelloAt = newInfo.LastHelloAt
-		if srv.addNodeInfoStmt != nil {
-			srv.addNodeInfo(&KnownNodeInfoWrapper{nodeid, newInfo})
-		}
-		if srv.GetRowIDStmt != nil {
-			if rowID := srv.getRowID(nodeid); rowID > 0 {
-				newInfo.RowID = rowID
-			}
-		}
-		srv.KnownNodeInfos.Infos()[id] = newInfo
-
-		// add the new node as a static node
-		srv.addNewStatic(id, newInfo)
-	} else {
-		currentInfo.Lock()
-		currentInfo.LastHelloAt = newInfo.LastHelloAt
-		currentInfo.RemotePort = newInfo.RemotePort
-		if infoChanged(currentInfo, newInfo) {
-			currentInfo.IP = newInfo.IP
-			currentInfo.TCPPort = newInfo.TCPPort
-			currentInfo.P2PVersion = p2pVersion
-			currentInfo.ClientId = clientId
-			currentInfo.Caps = caps
-			currentInfo.ListenPort = listenPort
-			if srv.addNodeInfoStmt != nil {
-				// TODO: check logic
-				// in-memory entry should keep the Ethereum Status info
-				// new entry to the mysql db should contain only the new address, DEVp2p info
-				// let Ethereum protocol update the Status info, if available.
-				srv.addNodeInfo(&KnownNodeInfoWrapper{nodeid, currentInfo})
-			}
-			if srv.GetRowIDStmt != nil {
-				if rowID := srv.getRowID(nodeid); rowID > 0 {
-					currentInfo.RowID = rowID
-				}
-			}
-			// if the node's listening port changed
-			// add it as a static node
-			if currentInfo.TCPPort != newInfo.TCPPort {
-				srv.addNewStatic(id, newInfo)
-			}
-		} else {
-			if srv.updateNodeInfoStmt != nil {
-				srv.updateNodeInfo(&KnownNodeInfoWrapper{nodeid, currentInfo})
-			}
-		}
-		currentInfo.Unlock()
-	}
-}
-
-func infoChanged(oldInfo *Info, newInfo *Info) bool {
-	return oldInfo.IP != newInfo.IP || oldInfo.TCPPort != newInfo.TCPPort || oldInfo.P2PVersion != newInfo.P2PVersion ||
-		oldInfo.ClientId != newInfo.ClientId || oldInfo.Caps != newInfo.Caps || oldInfo.ListenPort != newInfo.ListenPort
-}
-
-// During the initial node info loading process
-// if a node seems to be listening (ie TCPPort != 0)
-// add it as a static node
-func (srv *Server) addInitialStatic(id discover.NodeID, nodeInfo *Info) {
-	if nodeInfo.TCPPort != 0 {
-		var ip net.IP
-		if ip = net.ParseIP(nodeInfo.IP); ip == nil {
-			log.Debug("Failed to add node to initial StaticNodes list", "node", fmt.Sprintf("enode://%s@%s:%d", id.String(), nodeInfo.IP, nodeInfo.TCPPort), "err", "failed to parse ip")
-		} else {
-			log.Debug("Adding node to initial StaticNodes list", "node", fmt.Sprintf("enode://%s@%s:%d", id.String(), nodeInfo.IP, nodeInfo.TCPPort), "err", "failed to parse ip")
-			// Ensure the IP is 4 bytes long for IPv4 addresses.
-			if ipv4 := ip.To4(); ipv4 != nil {
-				ip = ipv4
-			}
-			srv.StaticNodes = append(srv.StaticNodes, discover.NewNode(id, ip, nodeInfo.TCPPort, nodeInfo.TCPPort))
-		}
-	}
-}
-
-// if a node seems to be listening (ie TCPPort != 0)
-// add it as a static node
-func (srv *Server) addNewStatic(id discover.NodeID, nodeInfo *Info) {
-	if nodeInfo.TCPPort != 0 {
-		var ip net.IP
-		if ip = net.ParseIP(nodeInfo.IP); ip == nil {
-			log.Debug("Failed to add static node", "node", fmt.Sprintf("enode://%s@%s:%d", id.String(), nodeInfo.IP, nodeInfo.TCPPort), "err", "failed to parse ip")
-		} else {
-			// Ensure the IP is 4 bytes long for IPv4 addresses.
-			if ipv4 := ip.To4(); ipv4 != nil {
-				ip = ipv4
-			}
-			srv.AddPeer(discover.NewNode(id, ip, nodeInfo.TCPPort, nodeInfo.TCPPort))
-		}
-	}
-}
-
-func (srv *Server) prepareAddNodeInfoStmt() {
-	fields := []string{"node_id", "ip", "tcp_port", "remote_port", "p2p_version", "client_id", "caps", "listen_port",
-		"first_hello_at", "last_hello_at"}
-
-	stmt := fmt.Sprintf(`INSERT INTO node_info (%s) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-		strings.Join(fields, ", "))
-	pStmt, err := srv.DB.Prepare(stmt)
-	if err != nil {
-		log.Debug("Failed to prepare AddNodeInfo sql statement", "err", err)
-	} else {
-		log.Trace("Prepared AddNodeInfo sql statement")
-		srv.addNodeInfoStmt = pStmt
-	}
-}
-
-func (srv *Server) prepareUpdateNodeInfoStmt() {
-	pStmt, err := srv.DB.Prepare("UPDATE node_info SET remote_port=?, last_hello_at=? WHERE id=?")
-
-	if err != nil {
-		log.Debug("Failed to prepare UpdateNodeInfo sql statement", "err", err)
-	} else {
-		log.Trace("Prepared UpdateNodeInfo sql statement")
-		srv.updateNodeInfoStmt = pStmt
-	}
-}
-
-func (srv *Server) prepareAddNodeMetaInfoStmt() {
-	var updateFields []string
-	fields := []string{"node_id", "hash", "dial_count", "accept_count", "too_many_peers_count"}
-	for _, f := range fields[2:] {
-		updateFields = append(updateFields, fmt.Sprintf("%s=%s+VALUES(%s)", f, f, f))
-	}
-	stmt := fmt.Sprintf(`INSERT INTO node_meta_info (%s) VALUES (?, ?, ?, ?, ?) ON DUPLICATE KEY UPDATE %s`,
-		strings.Join(fields, ", "), strings.Join(updateFields, ", "))
-	pStmt, err := srv.DB.Prepare(stmt)
-	if err != nil {
-		log.Debug("Failed to prepare AddNodeMetaInfo sql statement", "err", err)
-	} else {
-		log.Trace("Prepared AddNodeMetaInfo sql statement")
-		srv.addNodeMetaInfoStmt = pStmt
-	}
-}
-
-func (srv *Server) prepareGetRowID() {
-	pStmt, err := srv.DB.Prepare("SELECT MAX(id) FROM node_info WHERE node_id=?")
-	if err != nil {
-		log.Debug("Failed to prepare GetRowID sql statement", "err", err)
-	} else {
-		log.Trace("Prepared GetRowID sql statement")
-		srv.GetRowIDStmt = pStmt
-	}
-}
-
-func (srv *Server) addNodeInfo(newInfoWrapper *KnownNodeInfoWrapper) {
-	nodeid := newInfoWrapper.NodeId
-	newInfo := newInfoWrapper.Info
-	firstUnixTime := float64(newInfo.LastHelloAt.UnixNano()) / 1000000000
-	lastUnixTime := float64(newInfo.LastHelloAt.UnixNano()) / 1000000000
-	_, err := srv.addNodeInfoStmt.Exec(nodeid, newInfo.IP, newInfo.TCPPort, newInfo.RemotePort,
-		newInfo.P2PVersion, newInfo.ClientId, newInfo.Caps, newInfo.ListenPort, firstUnixTime, lastUnixTime)
-	if err != nil {
-		log.Debug("Failed to execute AddNodeInfo sql statement", "id", nodeid[:16], "newInfo", newInfo, "err", err)
-	} else {
-		log.Trace("Executed AddNodeInfo sql statement", "id", nodeid[:16], "newInfo", newInfo)
-	}
-}
-
-func (srv *Server) updateNodeInfo(newInfoWrapper *KnownNodeInfoWrapper) {
-	nodeid := newInfoWrapper.NodeId
-	newInfo := newInfoWrapper.Info
-	unixTime := float64(newInfo.LastHelloAt.UnixNano()) / 1000000000
-	_, err := srv.updateNodeInfoStmt.Exec(newInfo.RemotePort, unixTime, newInfo.RowID)
-	if err != nil {
-		log.Debug("Failed to execute UpdateNodeInfo sql statement", "id", nodeid[:16], "newInfo", newInfo, "err", err)
-	} else {
-		log.Trace("Executed UpdateNodeInfo sql statement", "id", nodeid[:16], "newInfo", newInfo)
-	}
-}
-
-func boolToInt(b bool) int {
-	if b {
-		return 1
-	}
-	return 0
-}
-
-func (srv *Server) addNodeMetaInfo(nodeid string, hash string, dial bool, accept bool, tooManyPeers bool) {
-	_, err := srv.addNodeMetaInfoStmt.Exec(nodeid, hash, boolToInt(dial), boolToInt(accept), boolToInt(tooManyPeers))
-	if err != nil {
-		log.Debug("Failed to execute AddNodeMetaNodeInfo sql statement", "id", nodeid, "dial", dial, "accept", accept, "tooManyPeers", tooManyPeers, "err", err)
-	} else {
-		log.Trace("Executed AddNodeMetaNodeInfo sql statement", "id", nodeid, "dial", dial, "accept", accept, "tooManyPeers", tooManyPeers)
-	}
-}
-
-func (srv *Server) getRowID(nodeid string) uint64 {
-	var rowID uint64
-	err := srv.GetRowIDStmt.QueryRow(nodeid).Scan(&rowID)
-	if err != nil {
-		log.Debug("Failed to execute GetRowID sql statement", "id", nodeid, "err", err)
-		return 0
-	} else {
-		log.Trace("Execute GetRowID sql statement", "id", nodeid, "rowid", rowID)
-		return rowID
-	}
-}
-
 // PeersInfo returns an array of metadata objects describing connected peers.
 func (srv *Server) PeersInfo() []*PeerInfo {
 	// Gather all the generic and sub-protocol specific infos
@@ -1413,106 +917,6 @@ func (srv *Server) PeersInfo() []*PeerInfo {
 	for i := 0; i < len(infos); i++ {
 		for j := i + 1; j < len(infos); j++ {
 			if infos[i].ID > infos[j].ID {
-				infos[i], infos[j] = infos[j], infos[i]
-			}
-		}
-	}
-	return infos
-}
-
-type knownNodeInfos struct {
-	mux   sync.RWMutex
-	infos map[discover.NodeID]*Info
-}
-
-func (k *knownNodeInfos) Lock() {
-	k.mux.Lock()
-}
-
-func (k *knownNodeInfos) Unlock() {
-	k.mux.Unlock()
-}
-
-func (k *knownNodeInfos) RLock() {
-	k.mux.RLock()
-}
-
-func (k *knownNodeInfos) RUnlock() {
-	k.mux.RUnlock()
-}
-
-func (k *knownNodeInfos) Infos() map[discover.NodeID]*Info {
-	return k.infos
-}
-
-// KnownNodeInfo represents a short summary of the information known about a known node.
-type Info struct {
-	mux sync.RWMutex
-
-	RowID           uint64     `json:"RowID"`                     // Most recent row ID
-	Keccak256Hash   string     `json:"keccak256Hash"`             // Keccak256 hash of node ID
-	LastConnectedAt *time.Time `json:"lastConnectedAt,omitempty"` // Last time the node was connected
-	IP              string     `json:"ip"`                        // IP address of the node
-	TCPPort         uint16     `json:"tcpPort"`                   // TCP listening port for RLPx
-	RemotePort      uint16     `json:"tcpPort"`                   // Remote TCP port of the most recent connection
-
-	// DEVp2p Hello info
-	P2PVersion   uint64     `json:"p2pVersion,omitempty"`   // DEVp2p protocol version
-	ClientId     string     `json:"clientId,omitempty"`     // Name of the node, including client type, version, OS, custom data
-	Caps         string     `json:"caps,omitempty"`         // Node's capabilities
-	ListenPort   uint16     `json:"listenPort,omitempty"`   // Listening port reported in the node's DEVp2p Hello
-	FirstHelloAt *time.Time `json:"firstHelloAt,omitempty"` // First time the node sent Hello
-	LastHelloAt  *time.Time `json:"lastHelloAt,omitempty"`  // Last time the node sent Hello
-
-	// Ethereum Status info
-	ProtocolVersion uint64     `json:"protocolVersion,omitempty"` // Ethereum sub-protocol version
-	NetworkId       uint64     `json:"networkId,omitempty"`       // Ethereum network ID
-	FirstReceivedTd *big.Int   `json:"firstReceivedTd,omitempty"` // First reported total difficulty of the node's blockchain
-	LastReceivedTd  *big.Int   `json:"lastReceivedTd,omitempty"`  // Last reported total difficulty of the node's blockchain
-	BestHash        string     `json:"bestHash,omitempty"`        // Hex string of SHA3 hash of the node's best owned block
-	GenesisHash     string     `json:"genesisHash,omitempty"`     // Hex string of SHA3 hash of the node's genesis block
-	FirstStatusAt   *time.Time `json:"firstStatusAt,omitempty"`   // First time the node sent Status
-	LastStatusAt    *time.Time `json:"lastStatusAt,omitempty"`    // Last time the node sent Status
-	DAOForkSupport  bool       `json:"daoForkSupport"`            // Whether the node supports or opposes the DAO hard-fork
-}
-
-func (k *Info) Lock() {
-	k.mux.Lock()
-}
-
-func (k *Info) Unlock() {
-	k.mux.Unlock()
-}
-
-func (k *Info) RLock() {
-	k.mux.RLock()
-}
-
-func (k *Info) RUnlock() {
-	k.mux.RUnlock()
-}
-
-type KnownNodeInfoWrapper struct {
-	NodeId string `json:"nodeid"` // Unique node identifier (also the encryption key)
-	Info   *Info  `json:"info"`
-}
-
-// NodeInfo gathers and returns a collection of metadata known about the host.
-func (srv *Server) KnownNodes() []*KnownNodeInfoWrapper {
-	srv.KnownNodeInfos.Lock()
-	defer srv.KnownNodeInfos.Unlock()
-	infos := make([]*KnownNodeInfoWrapper, 0, len(srv.KnownNodeInfos.Infos()))
-	for id, info := range srv.KnownNodeInfos.Infos() {
-		nodeInfo := &KnownNodeInfoWrapper{
-			id.String(),
-			info,
-		}
-		infos = append(infos, nodeInfo)
-	}
-	// Sort the result array alphabetically by node identifier
-	for i := 0; i < len(infos); i++ {
-		for j := i + 1; j < len(infos); j++ {
-			if infos[i].NodeId > infos[j].NodeId {
 				infos[i], infos[j] = infos[j], infos[i]
 			}
 		}
